fix incorrect song duration from jango connector
The jango connector was reporting very short durations on songs that
hadn't completely loaded yet. Because of this they would get scrobbled
right away.

To protect against this I've added a check to make sure the song is
fully loaded befor checking the duration.

It was also possible for the UI and sound manager to load before the
first_pic_change was triggered. Checks were added for the presence of
these objects before interacting with them. The connector will continue
to poll until they are loaded.