History of manual submit
Keeps a history of manually submitted tracks when the automagic
identification fails. Only saves the track to history when the user
enters a valid track in the popup and only lockups in history if the
initial validation fails.

Please note that this is based upon the assumption that the URL changes
with the track, which is the case for YouTube but most likely not the
case for other connectors. For this to work in all connectors, the
connector should send a unique ID to the scrobbler for each track and
the scrobbler should save, and do the lockup, on that ID. The popup
would also need to work on that ID instead of active tab's URL.