Fixes Rdio scrobbling issue
There's an issue where sometimes Rdio scrobbles a correct songname with an incorrect artist name. This seems to fix it, it looks like a bug to do with line 33 executing before parseInfo is finished, so I put it in a callback to force synchronicity. If there's a better way to do this, or if I've misdiagnosed the issue, I'd be happy to hear :)