Add v2 connector for plug.dj.
Steal the guts of the YouTube v2 connector to do song name trimming.

Unregister the old plug.dj connector, meaning this one is the only one
that shows up. It's marked as v2, and the old connector still exists.

Respect the snoozing feature in plug.dj to allow "pausing", at least of
the scrobbling. This doesn't seem to be reflected in the web-scrobbler
UI, but console messages about stopping and starting *are* output to the
console.

Use the "Two MutationObservers" pattern suggested by @alexesprit.