Account for TIDAL redesign.
Redesign seemed to roll in late last week.

data-test-id attributes are probably less than ideal but until I absolutely
have to use string matches on class attributes I will not do so.

I tried to follow the old structure as much as possible; Not totally sure why
the duration and album art are not restricted to the player selector.