Improve amazon connector (#2204)
The text that appears in Artist tag as featured artist, i.e. "Artist A feat. B" differs between the album pages and the library pages (where it is [feat. Artist B]) in square brackets.

This case normalizes the artist tag to not contain the square brackets if they are present.