Implement URL pattern match blacklisting feature
Extends the existing custom patterns feature by marking each pattern in
the customPatterns storage entry with a flag. When retrieving a
connector based on the URL, this flag is used to check if the URL should
be ignored.

The user-facing component includes a new body in the connector options
modal dialogue, mirroring the existing pattern input. For the prompt, a
new message is added which will need to be translated.