Select the album from the new table structure (#2874)
TIDAL recently switched from using <table>s in the credits to a raw-<div>-based solution. This exposes an edge case which I suppose was already there before this change: If you start by playing a song from an album, then use the Play Queue feature, the album link will stay in the lower left, resulting in bad albums on your scrobbles. I can't think of a good way to avoid this and it seems like a small edge case as-is, but without the credits to fix the data, it's a bigger problem. And in any case, this brings back the old behavior when playing from playists etc.