Select only album as album
Qobuz seems to have made a change
that results in a DOM that includes
multiple textual components
inside of the `.player__track-album`
element. The key components are
all included as `a` tags, and the second
`a` tag is always the album (with the first
being the artist, and the optional third
being the playlist if present).

This commit makes a change to
always select the second `a` tag within
`.player__track-album` to ensure only
the album is selected (till further changes
are made on the Qobuz end).

Resolves web-scrobbler/web-scrobbler#3095