Improve timer.ts robustness (#3752)
Since there is a replayDetectionTimer, So when the song played end, it would be set to replay, and call the `pause()` method.
So that `this.pausedOn` will not be null.
Then will play the next song on the playlist, and call the `start()` method.
However, when `update()` called, `this.pausedOn == null` is false, so `this.setTrigger` is not called. This make the 2nd and the rest in the playlist would NOT be scrobbled.