Don't log to console if debug logging is disabled
Avoids logging anything to the end user's browser console if they
haven't enabled debug logging.

This doesn't affect any of the logging to the extension's service worker
console or the console on the extension's options pages, just to the
user's main browser console when they're on a regular web page.

A couple of things that aren't great here:

This new `getGlobalOption()` function has a lot of duplicate code from
`getOption()` above it. I couldn't use `getOption()` as-is because a) we
don't have a connector to pass as the mandatory `connector` option in
this context, and b) `getOption()` calls `debugLog()` if the requested
option cannot be found, so if we called `debugLog()` from `getOption()`
this could cause an infinite loop.

The `Promise<unknown>` return type of `getGlobalOption()` is pretty
terrible - the promise should only ever resolve to `false` or to one of
the possible values of the `GlobalOptions` interface. I tried to make
that work but I have very minimal experience with TypeScript so I
couldn't work out the syntax.

There are no existing tests for `debugLog()` or `getOption()` here so I
haven't added any to cover this behaviour either.