Add a global option to enable debug logging
Some folks have asked for an option to disable debug logging from the
extension, to limit the amount of noise generated in the browser console
across every website. This adds an option (which defaults to disabling
the logging), though it isn't hooked up to anything yet.

I've added English translations for the option's UI to
`src/_locales/en/messages.json` since that seems to be the source file
that translations are automatically generated from, according to the
config in `.tx/config`.