Ensure debug log mesages are printed in order
re: https://github.com/web-scrobbler/web-scrobbler/pull/4067#discussion_r1310393413

Introducing an asynchronous check for user preferences within
`debugLog()` in a previous commit meant that the print order of log
messages was no longer guaranteed to be the same as the order in which
each call to `debugLog()` was made. Implementing a queue for log
messages ensures that they will be printed in the correct order.