Cache value of DEBUG_LOGGING_ENABLED option
re: PR comment https://github.com/web-scrobbler/web-scrobbler/pull/4067#discussion_r1317722578

It seems excessive to check the current value of this option every time
we want to print a log message, so we can instead check it once when
it's first needed and then cache the result. This means that if a user
changes the option then they'll need to refresh the page before it'll
take effect, but that feels like an acceptable trade-off.