Prepare for @typescript-eslint/eslint-plugin @ 8. (#4874)
* Ignore no-unused-expressions in tests.

The Vitest `expect` format doesn't play nice with
@typescript-eslint/no-unused-expressions. There is also a configuration way to
set this up, but it requires a migration to the new-style ESLint configuration,
mentioned in the warning which pops up at the moment as well.

```
Warning: React version not specified in eslint-plugin-react settings. See https://github.com/jsx-eslint/eslint-plugin-react#configuration .
```

* Convert expressions to if statements.

These triggered the no-unused-expressions rule.

* Wrap rejections in Errors.

* Wrap throws in Errors.

The ServiceCallResult enum suggests a more central place to introduce the
construction of Errors, but as I understand it using `new` there would affect
stack traces. Consider the obvious next refactor a TODO.

* Move typecheck into promise's `then`.

* as-force the rethrown error.