Fix album artist special case in TIDAL.
The "canonical" link is gone. Leaving the variable name the same, the same data
is stored in a `apple-itunes-app` meta tag.

The container for artist details is slightly different. See #5213.

The information about an album page is no longer a full URL, so we match on the
last two segments, which should be a literal "album" and then its ID.